Refactored DrinkDetails + created IngredientList component
