+++ src/components/DetailsPage/style.css
