+++ src/pages/DrinkDetails.js
@@ -1,7 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+
 import { fetchDrinkId } from '../services/fetchDrinks';
-import FoodCarousel from '../components/FoodDetails/FoodCarousel';
+import IngredientList from '../components/DetailsPage/IngredientList';
+import FoodCarousel from '../components/DetailsPage/FoodCarousel';
 
 const DrinkDetails = (props) => {
   const { id } = props.match.params;
@@ -17,24 +19,6 @@ const DrinkDetails = (props) => {
 
   const youtubeURL = () => String(singleDrink.strVideo).replace('watch?v=', 'embed/');
 
-  const filterIngredients = () =>
-    Object.keys(singleDrink)
-      .filter(
-        (key) => key.includes('Ingredient') && singleDrink[key] !== '' && singleDrink[key] !== null,
-      )
-      .reduce((object, key) => {
-        if (key.length === 14) {
-          return {
-            ...object,
-            [singleDrink[key]]: singleDrink[`strMeasure${key[key.length - 1]}`],
-          };
-        }
-        return {
-          ...object,
-          [singleDrink[key]]: singleDrink[`strMeasure${key[key.length - 2]}${key[key.length - 1]}`],
-        };
-      }, {});
-
   return loading && !singleDrink ? (
     <section>Loading...</section>
   ) : (
@@ -47,15 +31,7 @@ const DrinkDetails = (props) => {
         {singleDrink.strCategory}
         {singleDrink.strAlcoholic}
       </h3>
-      <ul>
-        {Object.entries(filterIngredients()).map((key, index) =>
-          key[1] === null ? (
-            <li data-testid={`${index}-ingredient-name-and-measure`}>{key[0]}</li>
-          ) : (
-            <li data-testid={`${index}-ingredient-name-and-measure`}>{`${key[0]} - ${key[1]}`}</li>
-          ),
-        )}
-      </ul>
+      <IngredientList singleItem={singleDrink} />
       <p data-testid="instructions">{singleDrink.strInstructions}</p>
       {singleDrink.strVideo ? (
         <iframe width="420" height="315" src={youtubeURL()} data-testid="video" />
